+++ crates/oxc_traverse/src/context/scoping.rs
@@ -1,0 +1,114 @@
+use oxc_semantic::{ScopeTree, SymbolTable};
+use oxc_syntax::scope::{ScopeFlags, ScopeId};
+
+use super::FinderRet;
+
+/// Traverse scope context.
+///
+/// Contains the scope tree and symbols table, and provides methods to access them.
+///
+/// `current_scope_id` is the ID of current scope during traversal.
+/// `walk_*` functions update this field when entering/exiting a scope.
+pub struct TraverseScoping {
+    scopes: ScopeTree,
+    symbols: SymbolTable,
+    current_scope_id: ScopeId,
+}
+
+// Public methods
+impl TraverseScoping {
+    /// Get current scope ID
+    #[inline]
+    pub fn current_scope_id(&self) -> ScopeId {
+        self.current_scope_id
+    }
+
+    /// Get scopes tree
+    #[inline]
+    pub fn scopes(&self) -> &ScopeTree {
+        &self.scopes
+    }
+
+    /// Get mutable scopes tree
+    #[inline]
+    pub fn scopes_mut(&mut self) -> &mut ScopeTree {
+        &mut self.scopes
+    }
+
+    /// Get symbols table
+    #[inline]
+    pub fn symbols(&self) -> &SymbolTable {
+        &self.symbols
+    }
+
+    /// Get mutable symbols table
+    #[inline]
+    pub fn symbols_mut(&mut self) -> &mut SymbolTable {
+        &mut self.symbols
+    }
+
+    /// Walk up trail of scopes to find a scope.
+    ///
+    /// `finder` is called with `ScopeId`.
+    ///
+    /// `finder` should return:
+    /// * `FinderRet::Found(value)` to stop walking and return `Some(value)`.
+    /// * `FinderRet::Stop` to stop walking and return `None`.
+    /// * `FinderRet::Continue` to continue walking up.
+    pub fn find_scope<F, O>(&self, finder: F) -> Option<O>
+    where
+        F: Fn(ScopeId) -> FinderRet<O>,
+    {
+        let mut scope_id = self.current_scope_id;
+        loop {
+            match finder(scope_id) {
+                FinderRet::Found(res) => return Some(res),
+                FinderRet::Stop => return None,
+                FinderRet::Continue => {}
+            }
+
+            if let Some(parent_scope_id) = self.scopes.get_parent_id(scope_id) {
+                scope_id = parent_scope_id;
+            } else {
+                return None;
+            }
+        }
+    }
+
+    /// Walk up trail of scopes to find a scope by checking `ScopeFlags`.
+    ///
+    /// `finder` is called with `ScopeFlags`.
+    ///
+    /// `finder` should return:
+    /// * `FinderRet::Found(value)` to stop walking and return `Some(value)`.
+    /// * `FinderRet::Stop` to stop walking and return `None`.
+    /// * `FinderRet::Continue` to continue walking up.
+    pub fn find_scope_by_flags<F, O>(&self, finder: F) -> Option<O>
+    where
+        F: Fn(ScopeFlags) -> FinderRet<O>,
+    {
+        self.find_scope(|scope_id| {
+            let flags = self.scopes.get_flags(scope_id);
+            finder(flags)
+        })
+    }
+}
+
+// Methods used internally within crate
+impl TraverseScoping {
+    /// Create new `TraverseScoping`
+    pub(super) fn new(scopes: ScopeTree, symbols: SymbolTable) -> Self {
+        Self {
+            scopes,
+            symbols,
+            // Dummy value. Immediately overwritten in `walk_program`.
+            current_scope_id: ScopeId::new(0),
+        }
+    }
+
+    /// Set current scope ID
+    #[inline]
+    pub(crate) fn set_current_scope_id(&mut self, scope_id: ScopeId) {
+        self.current_scope_id = scope_id;
+    }
+}
